fix: Use tuned files instead of using it as a module
The previous version of the kernel_settings role used `tuned` as
a python library, and had a `kernel_settings` module which was a
wrapper around this code.  However, `tuned` version 2.23 has changed
its internal API and it is no longer possible to use it as a python
library.  Instead, the kernel_settings role has been refactored to
read/write `tuned` config files, and let the `tuned` daemon manage
the settings.

In addition, `tuned` 2.23 changed the location of the profile directory,
so the kernel_settings role will now determine the location of the
profile directory depending on the `tuned` version.

The old `kernel_settings` module is removed, along with all of the
python unit testing code.

A new `kernel_settings_get_config` module has been created which will
simply parse and return the given config file as a `dict`.
