--- .github/config/ubuntu-requirements.txt
@@ -1,6 +1,0 @@
-libdbus-1-dev
-libgirepository1.0-dev
-python3-dev
-libssl-dev
-libcairo2-dev
-python2-dev
