+++ ansible_pytest_extra_requirements.txt
@@ -1,6 +1,3 @@
 # SPDX-License-Identifier: MIT
 
 # ansible and dependencies for all supported platforms
-ansible ; python_version > "2.6"
-idna<2.8 ; python_version < "2.7"
-PyYAML<5.1 ; python_version < "2.7"
