--- tests/run_ansible_for_testing.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-set -euxo pipefail
-
-ANSIBLE_DIR="$1" ; shift
-source "$ANSIBLE_DIR/hacking/env-setup"
-"$@"
