--- tests/tuned/etc/tuned/basic_settings/tuned.conf
@@ -1,24 +1,0 @@
-# tuned configuration
-#
-
-[main]
-summary=kernel settings
-
-[sysctl]
-fs.epoll.max_user_watches=785592
-fs.file-max=379724
-kernel.threads-max=29968
-vm.max_map_count=65530
-
-[sysfs]
-/sys/fs/selinux/avc/cache_threshold=256
-# spectre/meltdown - turn off fix for performance boost
-/sys/kernel/debug/x86/pti_enabled=0
-/sys/kernel/debug/x86/retp_enabled=0
-/sys/kernel/debug/x86/ibrs_enabled=0
-
-[vm]
-transparent_hugepages=never
-
-[bootloader]
-cmdline=spectre_v2=off nopti panic=10001 splash
