--- tests/tuned/etc/tuned/bogus/tuned.conf
@@ -1,2 +1,0 @@
-[bogus section]
-bogus=profile
