--- tests/tuned/etc/tuned/kernel_settings/tuned.conf
@@ -1,5 +1,0 @@
-# tuned configuration
-#
-
-[main]
-summary=kernel settings
