--- tests/tuned/etc/tuned/tuned-main.conf
@@ -1,42 +1,0 @@
-#daemon = 0
-#reapply_sysctl = 0
-# Global tuned configuration file.
-
-# Whether to use daemon. Without daemon it just applies tuning. It is
-# not recommended, because many functions don't work without daemon,
-# e.g. there will be no D-Bus, no rollback of settings, no hotplug,
-# no dynamic tuning, ...
-daemon = 1
-
-# Dynamicaly tune devices, if disabled only static tuning will be used.
-dynamic_tuning = 0
-
-# How long to sleep before checking for events (in seconds)
-# higher number means lower overhead but longer response time.
-sleep_interval = 1
-
-# Update interval for dynamic tunings (in seconds).
-# It must be multiply of the sleep_interval.
-update_interval = 10
-
-# Recommend functionality, if disabled "recommend" command will be not
-# available in CLI, daemon will not parse recommend.conf but will return
-# one hardcoded profile (by default "balanced").
-recommend_command = 1
-
-# Whether to reapply sysctl from the e.g /etc/sysctl.conf, /etc/sysctl.d, ...
-# If enabled these sysctls will be re-appliead after Tuned sysctls are
-# applied, i.e. Tuned sysctls will not override system sysctls.
-reapply_sysctl = 1
-
-# Default priority assigned to instances
-default_instance_priority = 0
-
-# Udev buffer size
-udev_buffer_size = 1MB
-
-# Log file count
-log_file_count = 2
-
-# Log file max size
-log_file_max_size = 1MB
