+++ tox.ini
@@ -1,15 +1,6 @@
 # SPDX-License-Identifier: MIT
 [lsr_config]
 lsr_enable = true
-commands_pre = bash -c '\
-    set -euxo pipefail; \
-    if [ -f {toxinidir}/tests/install_tuned_for_testing.sh ]; then \
-      {toxinidir}/tests/install_tuned_for_testing.sh; \
-    elif [ -f {toxinidir}/tests/kernel_settings/install_tuned_for_testing.sh ]; then \
-      {toxinidir}/tests/kernel_settings/install_tuned_for_testing.sh; \
-    else \
-      exit 1; \
-    fi'
 
 [lsr_ansible-lint]
 configfile = {toxinidir}/.ansible-lint
