--- tuned_requirements.txt
@@ -1,13 +1,0 @@
-# Note: this requirements file is used to specify what dependencies are
-# needed to perform unit testing of tuned and tuned related code.
-# To install these pip packages on Fedora, you may also need to install
-# some system RPM packages:
-# dnf -y install dbus-devel glib2-devel cairo-gobject-devel gobject-introspection-devel
-configobj
-decorator
-dbus-python
-PyGObject
-pyudev
-# note - this is *not* the procfs pip - this is not on pypi
-git+https://git.kernel.org/pub/scm/libs/python/python-linux-procfs/python-linux-procfs.git@v0.4.11 ; python_version < "3.0"
-git+https://git.kernel.org/pub/scm/libs/python/python-linux-procfs/python-linux-procfs.git@v0.6 ; python_version >= "3.0"
